first try at editing java code added some minor changes to commands logic — only the owner can control the bot now
